propagate ugrid logger (#4351)

+++ lib/iris/tests/unit/experimental/ugrid/cf/test_CFUGridConnectivityVariable.py
@@ -16,8 +16,7 @@ import iris.tests as tests  # isort:skip
 
 import numpy as np
 
-from iris.experimental.ugrid import logger
-from iris.experimental.ugrid.cf import CFUGridConnectivityVariable
+from iris.experimental.ugrid.cf import CFUGridConnectivityVariable, logger
 from iris.experimental.ugrid.mesh import Connectivity
 from iris.tests.unit.experimental.ugrid.cf.test_CFUGridReader import (
     netcdf_ugrid_variable,
@@ -26,7 +25,7 @@ from iris.tests.unit.experimental.ugrid.cf.test_CFUGridReader import (
 
 def named_variable(name):
     # Don't need to worry about dimensions or dtype for these tests.
-    return netcdf_ugrid_variable(name, "", np.int)
+    return netcdf_ugrid_variable(name, "", int)
 
 
 class TestIdentify(tests.IrisTest):
